zavork/requirements.txt
discord.py[voice]
yt-dlp
pydub
python-dotenv
PyNaCl
